AI-container-damage-detection-main/requirements.txt
# Core API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# UI Framework
streamlit==1.31.0

# HTTP Client
requests==2.31.0

# Data Processing
pandas==2.2.0
openpyxl==3.1.2

# PDF Generation
fpdf2==2.7.7

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.1

# Computer Vision
opencv-python-headless==4.9.0.80
Pillow==10.2.0
scikit-image==0.22.0

# Deep Learning
torch==2.2.0
torchvision==0.17.0
ultralytics==8.1.24

# Segmentation (SAM / Mask R-CNN)
segment-anything>=1.0
# detectron2  # Optional: Install separately for Mask R-CNN support

# ML Severity Model
scikit-learn>=1.4.0
joblib>=1.3.0

# OCR (PaddleOCR primary, EasyOCR fallback)
paddleocr>=2.7.0
paddlepaddle>=2.5.0
easyocr==1.7.1
pytesseract==0.3.10

# Geometry
shapely==2.0.2

# OpenAI API
openai>=1.30.0

# Utilities
numpy==1.26.4